% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mems_data.R
\name{convert_mems_data}
\alias{convert_mems_data}
\title{Convert data from raw MEMS data file into format readable by TBsim}
\usage{
convert_mems_data(filename = NULL, data = NULL, limit = NULL)
}
\arguments{
\item{filename}{filename to read data from (CSV)}

\item{data}{data.frame with MEMS data (either `filename` or `data` need to be specified)}
}
\description{
Convert data from raw MEMS data file into format readable by TBsim
}
