% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_default_mems_data.R
\name{copy_default_mems_data}
\alias{copy_default_mems_data}
\title{Copy the default MEMS data file to a user's folder}
\usage{
copy_default_mems_data(userId, base = "/data/tbsim")
}
\description{
Copy the default MEMS data file to a user's folder
}
