% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sim.R
\name{do_sim}
\alias{do_sim}
\title{Run simulation wrapper}
\usage{
do_sim(user = NULL, input = list(), userData = NULL, userDetails = NULL,
  jobscheduler = FALSE, description = "No description", nPatients = 1,
  regimen = NULL, memsFile = NULL, drugNames = c())
}
\description{
Run simulation wrapper
}
