% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_plots.R
\name{gather_plots}
\alias{gather_plots}
\title{Gather TBsim plots into an object for showing in Shiny}
\usage{
gather_plots(id, user, session, type = "population", quick_sim = NULL,
  drug_definitions = NULL, custom_drugs = NULL)
}
\description{
Gather TBsim plots into an object for showing in Shiny
}
