% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_results_list.R
\name{get_sim_results_list}
\alias{get_sim_results_list}
\title{Get list of simulation results}
\usage{
get_sim_results_list(folder = "/data/tbsim", user = "ronkeizer@gmail.com",
  pattern = "", read_output = FALSE, write_db = TRUE, type = "all")
}
\description{
Get list of simulation results
}
