% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_regimen.R
\name{load_regimen}
\alias{load_regimen}
\title{Load a regimen from a TBsim regimen file into
a table format for displaying in the app}
\usage{
load_regimen(regimen = NULL, file = "therapy.txt", folder = NULL,
  drugDefinitions = NULL)
}
\description{
Load a regimen from a TBsim regimen file into
a table format for displaying in the app
}
