% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_regimen_list_default.R
\name{load_regimen_list_default}
\alias{load_regimen_list_default}
\title{Update the cached drug regimen list with}
\usage{
load_regimen_list_default()
}
\description{
Update the cached drug regimen list with
}
