% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_saved_results.R
\name{load_saved_results}
\alias{load_saved_results}
\title{Load saved results}
\usage{
load_saved_results(id, user, session, type = "population", quick_sim = NULL,
  drug_definitions = NULL, custom_drugs = NULL)
}
\description{
Load saved results
}
