% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBactResSplit}
\alias{plotBactResSplit}
\title{Plot Bacterial resistance split by compartment}
\usage{
plotBactResSplit(res, custom_drugs = NULL, regimen_colors = NULL)
}
\description{
Plot Bacterial resistance split by compartment
}
