% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_settings_summary.R
\name{read_settings_summary}
\alias{read_settings_summary}
\title{Read settings and make summary}
\usage{
read_settings_summary(folder, settingsInfo)
}
\description{
Read settings and make summary
}
