% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen_table_to_TBsim_format.R
\name{regimen_table_to_TBsim_format}
\alias{regimen_table_to_TBsim_format}
\title{Convert Regimen Table to TBsim format}
\usage{
regimen_table_to_TBsim_format(dat)
}
\description{
Convert Regimen Table to TBsim format
}
