% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload_all_drug_definitions.R
\name{reload_all_drug_definitions}
\alias{reload_all_drug_definitions}
\title{Reload all drug definitions}
\usage{
reload_all_drug_definitions(drugNames, user)
}
\description{
Reload all drug definitions
}
