% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_new_regimen.R
\name{save_new_regimen}
\alias{save_new_regimen}
\title{Save regimen specified as table into TBsim format}
\usage{
save_new_regimen(reg, description, user)
}
\description{
Save regimen specified as table into TBsim format
}
