% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_wrapper_single.R
\name{sim_wrapper_single}
\alias{sim_wrapper_single}
\title{Wrapper for single subject simulation}
\usage{
sim_wrapper_single(user = NULL, input = input, userData = NULL,
  nPatients = 1, output, description = "No description", progress = NULL,
  regimen = NULL, drugNames = c(), memsFile = NULL)
}
\description{
Wrapper for single subject simulation
}
