% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_ago.R
\name{time_to_ago}
\alias{time_to_ago}
\title{Convert a time into how long "ago" it was}
\usage{
time_to_ago(dat, sort = TRUE, to_character = TRUE, add_ago = TRUE)
}
\description{
Convert a time into how long "ago" it was
}
