% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mems_list.R
\name{update_mems_list}
\alias{update_mems_list}
\title{Refresh MEMS files for Shiny app}
\usage{
update_mems_list(cache = NULL, user = NULL, base_path = "/data/tbsim")
}
\description{
Refresh MEMS files for Shiny app
}
